crud parcial desenhista

+++ editoragibi/src/main/java/com/grupo2/editoragibi/Service/BaseObjects/BaseEdicoesGibi.java
@@ -3,15 +3,15 @@ package com.grupo2.editoragibi.Service.BaseObjects;
 import java.time.LocalDate;
 
 public class BaseEdicoesGibi {
-    protected long edicaoGibi_id;
+    protected int edicaoGibi_id;
     protected int nroEdicao;
     protected LocalDate dataPub;
     protected boolean publicada;
     
-    public long getEdicaoGibi_id() {
+    public int getEdicaoGibi_id() {
         return edicaoGibi_id;
     }
-    public void setEdicaoGibi_id(long edicaoGibi_id) {
+    public void setEdicaoGibi_id(int edicaoGibi_id) {
         this.edicaoGibi_id = edicaoGibi_id;
     }
     public int getNroEdicao() {
